Added error pages for delete, update, view seller products. Also added error pages for the products that do not exist

+++ Views/Product/deleteSellerProduct.php
@@ -1,3 +1,10 @@
 <?php
-Product::deleteProduct($_GET['id']);
-header("Location: ?controller=product&action=sellerProduct");
+$product = new Product($_GET["id"]);
+if($product->getUserId() != $_SESSION["user_id"]) {
+    header("Location: /?controller=general&action=error");
+}
+else {
+    Product::deleteProduct($_GET['id']);
+    header("Location: ?controller=product&action=sellerProduct");
+}
+
